Adjust documentation, packages, and method names to be more exact.
Inspired by oracle's guide to javadoc comments,
https://www.oracle.com/technical-resources/articles/java/javadoc-tool.html
I decided to take another crack at clarifying some descriptions.

Additionally, since I've started routing for CookieClicker the
version has changed and I had to represent that in a package change.

Finally, the getEffect method of effects were named confusingly.
Since the interface is already named effect, it was not very
descriptive to query for an effect, when really we were querying
for a number (or factor to use somewhere).

I've added package documentation, and an overview.html to explain
some of the keywords I use with my documentation. I may need to
still link to this.