Add text version of features removed from this first release.
I decided to cut a method to seek to the next major moment of
simulation because I felt it could not be adequately explained and
tested for library users. Instead of forgetting it, I wanted to
quickly draft up this vision and put it in source code so I
wouldn't forget, and to maybe gauge interest.

The wording of this text document could have been considered
for a little longer, I'm committing it right after writing it
but I didn't want to stall myself out with something as small
as this.