Correct maven configuration, and downgrade required java version.
One issue I noted in my code is that I relied on non-LTS java
versions (>= 14) with some switch expressions, so when I started
choosing a version to land on in maven I decided to go to 11
and refactor my code to work with it.

An unrelated change, but a small correction,
I had the wrong package name.

I wonder if not owning this domain in the future: cookie.com
on the web will become a problem.