Adjust tests for refund transactions.
These tests ensured that the amount promised for a refund by
getTransactionalBuildingAmount was equal to the amount delivered
to the instance's bank account. The tests were flawed in that
the documentation for getTransactionalBuildingAmount reads that
the amount returned is negative to indicate refunds.

This means when calculating the difference in bank account
totals between the total after the refund and before the refund
we expect a positive value, so we must negate the predicted
building amount.