Rename purchasable to isPurchasable.
This change is mostly spurred on that I want to perfect my method
names, signatures, and documentation before I finally mark the
package as a major version. Then, hopefully, I'll stop screwing
around. This change isn't major since I am the only user of this
package right now, and for me it's a simple change with IntelliJ.