Apply small documentation improvements.
This was a final round up of changes I found in my classes
that I wanted to apply before I was complete.

I finished replacing X2 with the keyword double.
I found the last reference to the mouse, and replaced it just with
clicking. I'm okay with leaving everything else be.